Rule change: rollin-to-place-in-active-list
This rule says that new players shall be inserted relative to the
current list of active raiders. When there are n active, listed,
players in a raid, there are n+1 holes to which a new player may be
inserted. The new player shall roll from 1 to n+1 and will then occupy
that hole. The insertion with respect to the entire list will occur
either immediately before the first active raider, or immediately
after the appropriate active raider.